Move prompts to tables. Let strategies have their own configuration. (#17)
# Overview

This PR primarily moves a number of things that had been hardcoded out
into the database so they can be reconfigured without redeploying.

We introduce 4 new tables:
* `PromptTemplate` - A template for a text prompt to be sent to an
inference model, with support for template variables and control
structures in Jinja2 format.

* `InferenceModel` - A predictive model such as an LLM, image generation
model, or image analysis model. At this time all such models are
provided by external providers such as OpenAI, Stability, HuggingFace,
or Microsoft Azure.

* `ModelConfig` - A set of parameters and optional `PromptTemplate` for
use with a particular `InferenceModel`.

* `StrategyConfig` - A configuration for a story strategy, including
parameters, model configs for text->text and text->image, and an
optional "seed prompt" to help begin a story.

# Details

The Calliope project stagnated for a few months, and this PR with it. In
addition, I have completely changed development environments and am
somewhat lucky to be able to recover certain of these changes at all. As
a consequence, it has been a challenge to return to the PR and come back
up to speed on everything needed to get it ready for merge. Now, I think
I'm finally there.