Begin using LangChain. (#22)
This PR adds the LangChain library and begins using it for LLM
inferences. In support of this, the `inference` module is subdivided
into lots of smaller supporting modules.

Model engines currently able to use LangChain are HuggingFace and
OpenAI.

Because LangChain works only with text as both input and output, we need
to maintain the classic inference framework to work with images for
image->text and text->image references.