Convert Clio to TypeScript. (#29)
This PR converts the Clio app from JavaScript/JSX to TypeScript/TSX.
This is functionally analogous to adding type annotations and mypy to a
Python program. The result is cleaner, safer, more demonstrably correct
code. Indeed, the process enabled me to identify and correct some
existing bugs having to do with menu selection.