Adding location metadata. (#40)
# Overview

This PR adds location metadata based on the caller's IP address. This
includes things like city, country, time zone, local time, current
weather conditions, etc.

In addition:

* A new strategy, "lavender", takes advantage of the new metadata, and
is able to accept a JSON response from the LLM.
* The new metadata, along with other useful new information about model
and prompt selection, is included in the debug data recorded during
frame generation.
 
Hence, starting now, we can tell where users were when story frames are
generated. Calliope is free to use this to influence the story, and we
can also see this when reviewing stories in Thoth or Calliope Admin.