New storytellers Tamarisk, Narcissus, and Fern. New thinking about image analysis and story maintenance. (#48)
# Overview

The main points are:
* A few new storytellers (see below).
* Deeper image analysis thanks to gpt-4o.
* A few bug fixes in Clio.
* Start moving away from prompt templates in the database toward just
doing them straight in the code.
* Introduce `instructor`, and start using Pydantic models to formalize
the LLM interface.
* Some package updates.

## New Storytellers

### Tamarisk
Uses gpt-neo-2.7B in stream-of-consciousness mode, with slight editorial
gloss by gpt-4o. Meant to be fun and kind of random.

### Narcissus
A distorting mirror. Generates an image based on the image analysis. No
text output.

### Fern
The new flagship storyteller. Maintains a clear notion of story state,
with a genre, a concept, a cast of characters, specific settings, etc.
Far from perfect, but an major step forward.